requirements.txt
asgiref==3.5.0; python_version >= "3.8"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
cryptography==36.0.2; python_version >= "3.6"
deprecated==1.2.13; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
dj-database-url==0.5.0
django-cors-headers==3.11.0; python_version >= "3.7"
django-filter==21.1; python_version >= "3.6"
django-oauth-toolkit==1.7.1
django-rest-swagger==2.2.0
django==4.0.3; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
drf-yasg==1.20.0; python_version >= "3.6"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-metadata==4.11.3; python_version < "3.10" and python_version >= "3.7"
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_version >= "3.6"
jinja2==3.1.2; python_version >= "3.7"
jwcrypto==1.0
mailchimp-transactional==1.0.47
markdown==3.3.6; python_version >= "3.6"
markupsafe==2.1.1; python_version >= "3.7"
oauthlib==3.2.0; python_version >= "3.6"
openapi-codec==1.3.2
packaging==21.3; python_version >= "3.6"
pillow==9.1.0; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pytz==2022.1; python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.6"
ruamel.yaml==0.17.21; python_version >= "3.6"
simplejson==3.17.6; python_version >= "2.5" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
wrapt==1.14.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
zipp==3.8.0; python_version < "3.10" and python_version >= "3.7"
gunicorn==20.1.0
whitenoise==6.1.0
sendgrid==6.9.7
python-http-client==3.3.7
starkbank-ecdsa==2.0.3